Responsive Design and SEO Improvements
The embedded Google Maps were not displaying properly on mobile. Additional styling was added in an attempt to fix this. The meta tags on each page were refined. This includes reducing the number of keywords and adding the appropriate canonical links for each page. Also updated the favicon to be a pizza, rather than just "PG".